requirements.txt
Flask
dotenv
gunicorn
requests
azure.monitor.opentelemetry